nsproject/requirement.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
djangorestframework==3.14.0
drf-spectacular==0.26.1
inflection==0.5.1
jsonschema==4.17.3
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
